remove gateway.
